+++ src/memex_logging/common/dao/common.py
@@ -26,7 +26,7 @@ from memex_logging.common.utils import Utils
 logger = logging.getLogger("logger.common.dao.common")
 
 
-class EntryNotFound(Exception):
+class DocumentNotFound(Exception):
     pass
 
 
@@ -36,17 +36,16 @@ class CommonDao:
         self._es = es
         self._base_index = base_index
 
-    def _generate_index(self, project: Optional[str] = None, dt: Optional[datetime] = None) -> str:
+    def _generate_index(self, dt: Optional[datetime] = None) -> str:
         """
-        Generate the Elasticsearch index associated to the message, the format is `data_type-project-%Y-%m-%d`.
+        Generate the Elasticsearch index associated to the message, the format is `data_type-%Y-%m-%d`.
 
-        :param Optional[str] project: the project associated to the message
         :param Optional[datetime] dt: the datetime of the message
         :return: the generated Elasticsearch index
         :raise ValueError: when there is a datetime but not a project
         """
 
-        return Utils.generate_index(self._base_index, project=project, dt=dt)
+        return Utils.generate_index(self._base_index, dt=dt)
 
     @staticmethod
     def _build_query_by_id(trace_id: str) -> dict:
@@ -105,7 +104,7 @@ class CommonDao:
         )
         return query
 
-    def add(self, index: str, object_repr: dict, doc_type: str = "_doc") -> str:
+    def _add_document(self, index: str, object_repr: dict, doc_type: str = "_doc") -> str:
         """
         Add a document to Elasticsearch
 
@@ -118,7 +117,19 @@ class CommonDao:
         query = self._es.index(index=index, body=object_repr, doc_type=doc_type)
         return query["_id"]
 
-    def get(self, index: str, query: dict) -> Tuple[dict, str]:
+    def _update_document(self, index: str, trace_id: str, object_repr: dict, doc_type: str = "_doc") -> None:
+        """
+        Add a document to Elasticsearch
+
+        :param str index: the index where to add the document
+        :param str trace_id: the id of the document
+        :param dict object_repr: the document to add
+        :param str doc_type: the type of the document
+        """
+
+        self._es.index(index=index, id=trace_id, doc_type=doc_type, body=object_repr)
+
+    def _get_document(self, index: str, query: dict) -> Tuple[dict, str, str, str]:
         """
         Retrieve a document from Elasticsearch
 
@@ -131,11 +142,13 @@ class CommonDao:
         response = self._es.search(index=index, body=query)
         if len(response['hits']['hits']) == 0:
             logger.debug("Could not find any document")
-            raise EntryNotFound("Could not find any document")
-        else:
-            return response['hits']['hits'][0]['_source'], response['hits']['hits'][0]['_id']
+            raise DocumentNotFound(f"No document was found")
+        elif len(response['hits']['hits']) > 1:
+            logger.warning(f"More than one document was found")
 
-    def delete(self, index: str, query: dict) -> None:
+        return response['hits']['hits'][0]['_source'], response['hits']['hits'][0]['_id'], response['hits']['hits'][0]['_index'], response['hits']['hits'][0]['_type']
+
+    def _delete_document(self, index: str, query: dict) -> None:
         """
         Delete a document from Elasticsearch
 
@@ -145,13 +158,13 @@ class CommonDao:
 
         self._es.delete_by_query(index=index, body=query)
 
-    def search(self, index: str, query: dict) -> List[dict]:
+    def _search_documents(self, index: str, query: dict) -> List[dict]:
         """
         Search documents in Elasticsearch
 
-        :param str index: the index from which to search the documents
-        :param dict query: the query for searching the documents
-        :return: a list containing the representation of objects
+        :param str index: the target index for the search
+        :param dict query: the search query
+        :return: the list of documents matching the query
         """
 
         response = self._es.search(index=index, body=query)