+++ test/unit/memex_logging/common_test/mock/daos.py
@@ -19,8 +19,10 @@ from typing import Optional, List, Tuple
 
 from elasticsearch import Elasticsearch
 
+from memex_logging.common.dao.analytic import AnalyticDao
 from memex_logging.common.dao.collector import DaoCollector
 from memex_logging.common.dao.message import MessageDao
+from memex_logging.common.model.analytic.analytic import Analytic
 from memex_logging.common.model.message import Message
 
 
@@ -29,20 +31,41 @@ class MockMessageDao(MessageDao):
     A mock for the message dao
     """
 
-    def add_messages(self, message: Message, doc_type: str = "_doc") -> str:
+    def add(self, message: Message, doc_type: str = "_doc") -> str:
         pass
 
-    def get_message(self, project: Optional[str] = None, message_id: Optional[str] = None,
-                    user_id: Optional[str] = None, trace_id: Optional[str] = None) -> Tuple[Message, str]:
+    def get(self, project: Optional[str] = None, message_id: Optional[str] = None,
+            user_id: Optional[str] = None, trace_id: Optional[str] = None) -> Tuple[Message, str]:
         pass
 
-    def delete_message(self, project: Optional[str] = None, message_id: Optional[str] = None,
-                       user_id: Optional[str] = None, trace_id: Optional[str] = None) -> None:
+    def delete(self, project: Optional[str] = None, message_id: Optional[str] = None,
+               user_id: Optional[str] = None, trace_id: Optional[str] = None) -> None:
         pass
 
-    def search_messages(self, project: str, from_time: datetime, to_time: datetime, max_size: int,
-                        user_id: Optional[str] = None, channel: Optional[str] = None,
-                        message_type: Optional[str] = None) -> List[Message]:
+    def search(self, project: str, from_time: datetime, to_time: datetime, max_size: int,
+               user_id: Optional[str] = None, channel: Optional[str] = None,
+               message_type: Optional[str] = None) -> List[Message]:
+        pass
+
+
+class MockAnalyticDao(AnalyticDao):
+    """
+    A mock for the analytic dao
+    """
+
+    def add(self, analytic: Analytic, doc_type: str = "_doc") -> None:
+        pass
+
+    def update(self, index: str, trace_id: str, analytic: Analytic, doc_type: str = "_doc") -> None:
+        pass
+
+    def get(self, analytic_id: str) -> Analytic:
+        pass
+
+    def get_with_additional_information(self, analytic_id: str) -> Tuple[Analytic, str, str, str]:
+        pass
+
+    def delete(self, analytic_id: str) -> None:
         pass
 
 
@@ -53,6 +76,8 @@ class MockDaoCollectorBuilder:
 
     @staticmethod
     def build_mock_daos() -> DaoCollector:
+        es = Elasticsearch()
         return DaoCollector(
-            MockMessageDao(Elasticsearch())
+            MockMessageDao(es),
+            MockAnalyticDao(es)
         )